@@ -1,6 +1,6 @@
-const validator = require('../utils/validator');
+import validator = require('../utils/validator');
 const { isAlpha, isDate, isNumeric } = require('validator');
-const AppError = require('../utils/appError');
+import AppError = require('../utils/appError');
 const { RESULTS_SEARCH_LIMIT, DEFAULT_SORT_FIELD } = require('../config');
 
 const PARAMS_TO_FILTER = [
